requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in local.in, then run this to update:
#
#    pip-compile local.in
#
argparse==1.4.0           # via progressbar2
click==6.3                # via pip-tools
decorator==4.0.9          # via ipython, traitlets
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-dotenv==1.4.1
django-extensions==1.6.1
django-filter==0.13.0
django-grappelli==2.8.1
django-pipeline==1.6.7
django-stdimage==2.3.2
django==1.9.4
djangorestframework==3.3.2
first==2.0.1              # via pip-tools
ipython-genutils==0.1.0   # via traitlets
ipython==4.1.2
markdown==2.6.5
path.py==8.1.2            # via pickleshare
pexpect==4.0.1            # via ipython
pickleshare==0.6          # via ipython
pillow==3.1.1             # via django-stdimage
pip-tools==1.6
progressbar2==3.6.2       # via django-stdimage
ptyprocess==0.5.1         # via pexpect
pygments==2.1.3
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via django-extensions, pip-tools
sqlparse==0.1.19          # via django-debug-toolbar
traitlets==4.1.0          # via ipython
Unipath==1.1

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via ipython
